move old fixtures to the module in which they are used to prevent relative imports. Added explanation of B904 needing an explicit exception to tox.ini
