+++ tox.ini
@@ -19,6 +19,7 @@ ignore=
 
 per-file-ignores=
     ; TYPE_CHECKING block suggestions
+    ; B904 needs explicitly excepting because it's added in extend-select
     tests/*: A, B, B904, C, E731, TC, ISC, N, SIM
 
     ; for clarity we don't merge 'with Conf():' context trees
